🛂 return JWT on /api/auth/login

+++ advent_calendar_backend/src/main/java/de/iks/advent_calendar_backend/security/JwtIssuer.java
@@ -11,7 +11,7 @@ import java.util.List;
 
 @Component
 public class JwtIssuer {
-	public String issue(int userId, String username, List<String> roles) {
+	public String issue(Long userId, String username, List<String> roles) {
 		return JWT.create()
 				.withSubject(String.valueOf(userId))
 				.withExpiresAt(Instant.now().plus(Duration.of(1, ChronoUnit.DAYS)))
